refactor: 404 page
