+++ pages/404/index.tsx
@@ -1,0 +1,5 @@
+import Page404 from "@/Views/Page404";
+
+export default function Page404Page() {
+    return <Page404 />;
+}
